Stage and Scene refactor
In order to organize things better, some new changes. Firstly, scenes should
take a stage parameter when created, which makes the stage not need to pass
itself to various functions. All of the places in a scene that used to take a
stage parameter now use the member version instead.

Secondly, Scenes have grown the ability to register actors, which means that
the Scene versions of update and render can loop through all actors and
update/render them, which simplifies things a bit.

Note that the code renders the actors in a specific order, so if you want to
do things with a background you need to draw that before calling into the
base class, and you need to register your actors (currently) in the order you
want them rendered.