Entity properties
Include the ability for entities to have properties. This is implemented by
having the base entity class contain a properties dictionary (may be empty).
The constructor will copy all properties from this.defaultProperties to the
properties, if such an object exists, to provide default values where sensible.

A subclass of entity can set this.defaultProperties before invoking the
constructor.

There is a method call that the base class constructor calls to allow the
instance to validate its properties.