Entities carry a stage
We want the entities to carry the stage object so that they can do things
like trigger scene changes. To this end, Actor and Entity both take an
optional stage parameter.

For the purposes of carrying entities with map data, the idea is that the
Level class takes a stage and it will set that stage into all entities in the
map data if they don't already have one.

This allows the entity classes that are a part of the level data to be
created beforehand and still have the stage.