Doors no longer close when blocked
Modify the door entities so that when they are toggling their state to close
themselves, they will not close if they have access to the current scene, and
that scene indicates that there is an actor standing in the door.

In the case of a manual toggle, the toggle is lost. In an automatic toggle it
is delayed for as long as the door is blocked.