New entity: button
Not finished yet, but visually you can place a button on a tile oriented on
any of the four faces, and it can appear pressed or released.

The next step will be to make the action button activate the button if it is
not already pressed, and have it trigger things like doors.