Level API change
The level now creates a cross reference of all entities according to their ID
values for fast lookup.

There is now a method that will look up entities by their IDs, which takes
either a single string or an array of strings and returns an array of all
matching entities.

There is also a new helper method which will use the new method to invoke the
trigger method on all entities with a matching ID value.