Allow entity triggers
ChronoEntity now supports an optional trigger property, which can be either a
single string ID of another entity or an array of strings.

There is also a helper method in the class that will collect the current
scene, get the current level from the current scene, then get that scene to
tell it all of the entities with IDs that match its trigger list and trigger
them, specifying itself as the activator.

This common logic can be used to allow any entity to trigger another entity
with a single line of code.