Level Swap
Now there are a bunch of buttons and doors to make the sample level more
interesting.

In particular this shows how a button can revert itself back to an unpressed
state if configured to do so, how a button can trigger a door, and how a
button can trigger another pressed button in order to have it re-enable itself.