Actor Z-Order for rendering
Actors (and thus Entity and ChronoEntity) now have a Z-Order value that
describes their relative layering order. The Scene class now has a function
to sort all of the registered actors according to their current zOrder value,
so that when they are rendered, they get displayed in an appropriate order.