Debug changes
There is now a F1 debug key that will display complete entity information for
all entities at the current debug location (where the mouse is currently
positioned on the canvas).

In addition to this, if any of those entities contain triggers, arrows will
be rendered that indicate the locations of the entities that will be
triggered, allowing for visual confirmation of how the level works.