Start of guard movement/patrols
This allows for guards to be able to patrol by providing them with a
property that gives the waypoint locations.

Validation can also be done that all of the patrol points given exist and are
waypoints.

Still to do: validate that the waypoints are straight horizontal or vertical
from each other so the guard can move.