Enhance step function
The step method in Actor is enhanced to take a reference to the level that
contains the actor in question. This is inherited by entities. The Level
class passes itself in the stepAllEntities() method.

This allows actors to make decisions during their step phase based on the
rest of the level.