Guard Intelligence Increase
Guards will now stop their patrol if it takes them into level geometry or out
of the world entirely, instead of just going right through.

The patrol is stopped entirely as if it is not set in this case, because the
guard can't possibly continue if level geometry is at fault.