Guard Intelligence Increase #2
Guards now pause their patrol if their route is blocked by an entity that
blocks actor movement. This is not a permanent block and they will try again
on the next turn to move again.

Additionally, when blocked by entities that they know can potentially be
bypassed (so far, only a door), they will attempt to trigger the door to open
it so they can go through.

Guards will now stop their patrol if it takes them into level geometry or out
of the world entirely, instead of just going right through.

The patrol is stopped entirely as if it is not set in this case, because the
guard can't possibly continue if level geometry is at fault.