Refactor Level (multiple steps needed)
We alter how the level data is returned from the global accessor so that it
takes a stage, allowing it to pass that to the LevelData constructor, so that
the LevelData can give it to the entities.

The code that was shifted over to the LevelData is also removed from here now.