Refactor Level
The API for getting at entities and checking for blocking is now X,Y and
Point compatible for robustness.

This is implements as the point version invoking the x,y version for speed
since the operations might get invoked a lot and making an extra point a
bunch of times (e.g. once for every tile on the screen every frame just to
determine what to render) seems like a Bad Move (tm).