Enhance trigger property
Previous the trigger property could be either a string or an array of strings.

Now the code in ChronoEntity checks to see if the property value is a string
and turns it into an array, and then validates that if the (optional)
property exists, it is an array.

This simplifies code elsewhere that otherwise needs to keep track of wether
there is one point or several.