Add Entity Interactivity
Now ChronoEntity can be queried about whether or not the player can interact
with an entity. The default is no, but buttons are an example of an entity
that can be interacted with.

The code that handles the interaction key now uses this to filter down the
list of entities that can be activated. If its empty, nothing happens,
including taking a turn.

if the Player wants to burn time, they have to use the wait option.