New ChronoEntity position API
ChronoEntity now maintains two positions, normal position in world
coordinates for rendering, and a map position for other uses.

There is a new API that allows both positions to be set based on setting the
position one way or the other.

This requires that all position changes now happen through the API and not
just on the fly by modifying the instance variable. That part comes next.