Rendering Tweaks
Some of the entity rendering code was taking liberties by assuming pixel
offsets for things, which mess things up if the tile size is not exactly 32.

Now everything that renders is based off of the tile size by calculating
percentages of the tile size when offsets are required.

This was done by changing pure pixel offsets into percentages of the tile
size, with rounding to ensure that things remain as centered as possible.