New ChronoEntity rendering API
In order to simplify an upcoming change in which entities have a facing, and
in particular how their rendering will change as a result, it is desirable to
have the canvas rotated so that a single rendering operation will draw the
entity as appropriate. This also requires the canvas origin to be at the
center of the map tile (since that is what the rotation is relative to).

ChronoEntity now has a couple of simple API functions that will set up this
translation and rotation using internal properties, and it now also does its
default rendering based on that new scheme.