ChronoEntity facing API
Add a new API (kinda untested ATM) that allows you to set the facing based on
either a numeric value or a string. The facing property now uses this
function to convert.

The function (in theory) will make sure that the number that you give it is
constrained to the closest multiple of 90 that we support.