Enhance facing API
There are now utility methods for determining what angle you would end up at
if you needed to turn to face a new direction, with the caveat that you only
want to turn 90 degrees at a time.

When a turn of 180 needs to be done, this is done by turning 90 degrees
either left or right, which is based on the handedness property on the entity
that is moving. This defaults to right handed but can be set to left handed
if desired (some guards would thus look in different directions when patrolling.

The player defaults to right handed turns because that's what the default is,
but the property could be set if desired.