New entity interaction API
Now you can ask an entity if it can be interacted with before you trigger it
to see if it wants to allow you to interact with it or not. Note that this
doesn't meant that it won't respond if you DO trigger it.

This is now used to ensure that a button can't be interacted with if you have
your back to it.