New button type: panel
The button class supports a new visual style, panel. This is signalled via a
property, where the default (false) causes it to render as a button, and
otherwise it renders as a panel.

A panel is a small gray rectangle with a red (released) or green (pressed)
light on it. In all other regards, it acts like a button usually acts.