Interaction with panel buttons
Buttons that are panels are meant to be interacted with from an adjacent tile
but need you to be facing 180 degrees from them, because you need to fully
engage the panel to operate it.

Buttons on the other hand can be pressed while facing or when you are
standing with it on your left or right (basically you just can't kick it or
back into it to trigger it), but you have to be on the same tile.