Interaction refactor
Split the code that finds entities to interact with into its own function. It
checks first the tile that the player is on, and then the tile in front of,
to the right and to the left (although the order will be left then right if
the handedness is left for the player; clever).

After this, the entities found are filtered to only those that want to
interact with the player, and the list is then returned.

As a simple test, the level has been extended so as to have a couple of doors
with mutual panels that the player needs to manipulate to get through.