Allow vision cones to be initially generated
When the game scene becomes active, we now prompt all guards to recreate
their vision cones. This is the first opportunity that they have to do so,
because until the game scene is the current scene, the guards have no access
to the level data they need to perform this task.