Fix scope leak of map position
The specialization of ChronoEntity to have a map position has been pushed up
to Actor in general, as anything might want to be aligned to the grid of
tiles or have a position inferred in that manner.

This allows the code in Level to use map position to get entities at a
particular location using map position without it being a scope leak from the
engine in general.