API rename
Pursuant to the previous check in, the APIs in level that are responsible
for checking for entities at a map location now make it clear that they are
searching in tile coordinates.

There is currently no analog for actual world coordinates, since the current
game does not use them and it's not immediately useful to find entities at an
exact pixel perfect location.

As such I'll let the natural genesis of things play out in determining how
that API should work.